wz/requirements.txt
# 微信公众号文章抓取工具依赖
requests>=2.28.1
lxml>=4.9.1
tqdm>=4.64.1
python-dateutil>=2.8.2
PyMySQL>=1.0.2

# Flask Web界面依赖
flask>=2.2.3
apscheduler>=3.10.1
python-dotenv>=1.0.0
werkzeug>=2.2.3

# DrissionPage (用于微信公众号登录认证)
DrissionPage>=4.0.0 
# 文章内容采集配置简易指南
# 文章内容采集配置简易指南
trafilatura
newspaper3k